List type matching for inspiralfuns
The input type checking is expecting these variables to be of the type list, but whenever np.zeros is involved they would actually be output as numpy.ndarray. This makes sure the output type is actually what the input is expecting.
The alternative would be to change the input type checking to expect numpy.ndarray instead of list, but this is probably more versatile for users given these are all one-dimensional series of numbers anyway.